refactor: handler package to http package
