+++ http/ApiHandler.go
@@ -1,7 +1,6 @@
-package handler
+package http
 
 import (
-	ghttp "go-backend/http"
 	"net/http"
 )
 
@@ -10,7 +9,7 @@ type ApiHandler func(w http.ResponseWriter, r *http.Request) error
 func (ah ApiHandler) serve(w http.ResponseWriter, r *http.Request) {
 	if err := ah(w, r); err != nil {
 		// handling error
-		ghttp.HandleError(err, w, r)
+		HandleError(err, w, r)
 	}
 }
